requirements.txt
Flask==3.0.0
Werkzeug==3.0.1
qrcode==7.4.2
python-dotenv==1.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Frozen-Flask==0.18
gunicorn==21.2.0
netlify_lambda_wsgi==0.1.5
Pillow==10.1.0
SQLAlchemy==2.0.23
Jinja2==3.1.2
itsdangerous==2.1.2
click==8.1.7
markupsafe==2.1.3
pytest==7.4.3
pytest-flask==1.3.0
